+++ nPBRT/nPBRT/Core/Geometry/Geometry.cs
@@ -8,12 +8,12 @@ namespace nPBRT.Core.Geometry
 {
     public class Geometry
     {
-        public static float Dot(Vector v1, Vector v2)
+        public static double Dot(Vector v1, Vector v2)
         {
             if (v1.HasNaNs() || v2.HasNaNs()) throw new InvalidOperationException();
             return v1.x * v2.x + v1.y * v2.y + v1.z * v2.z;
         }
-        public static float AbsDot(Vector v1, Vector v2)
+        public static double AbsDot(Vector v1, Vector v2)
         {
             if (v1.HasNaNs() || v2.HasNaNs()) throw new InvalidOperationException();
             return Math.Abs(Dot(v1, v2));
@@ -35,74 +35,74 @@ namespace nPBRT.Core.Geometry
         {
             if (Math.Abs(v1.x) > Math.Abs(v1.y))
             {
-                float invLen = 1.0f / (float)Math.Sqrt(v1.x * v1.x + v1.z * v1.z);
-                v2 = new Vector(-v1.z * invLen, 0.0f, v1.x * invLen);
+                double invLen = 1.0d / Math.Sqrt(v1.x * v1.x + v1.z * v1.z);
+                v2 = new Vector(-v1.z * invLen, 0.0d, v1.x * invLen);
             }
             else
             {
-                float invLen = 1.0f / (float)Math.Sqrt(v1.y * v1.y + v1.z * v1.z);
-                v2 = new Vector(0.0f, v1.z * invLen, -v1.y * invLen);
+                double invLen = 1.0d / Math.Sqrt(v1.y * v1.y + v1.z * v1.z);
+                v2 = new Vector(0.0d, v1.z * invLen, -v1.y * invLen);
             }
             v3 = Cross(v1, v2);
         }
 
-        public static float Distance(Point p1, Point p2)
+        public static double Distance(Point p1, Point p2)
         {
             return (p1 - p2).Length();
         }
-        public static float DistanceSquared(Point p1, Point p2)
+        public static double DistanceSquared(Point p1, Point p2)
         {
             return (p1 - p2).LengthSquared();
         }
 
-        public static float Dot(Normal n, Vector v)
+        public static double Dot(Normal n, Vector v)
         {
             if (n.HasNaNs() || v.HasNaNs()) throw new InvalidOperationException();
             return n.x * v.x + n.y * v.y + n.z * v.z;
         }
-        public static float Dot(Vector v, Normal n)
+        public static double Dot(Vector v, Normal n)
         {
             if (v.HasNaNs() || n.HasNaNs()) throw new InvalidOperationException();
             return v.x * n.x + v.y * n.y + v.z * n.z;
         }
-        public static float Dot(Normal n1, Normal n2)
+        public static double Dot(Normal n1, Normal n2)
         {
             if (n1.HasNaNs() || n2.HasNaNs()) throw new InvalidOperationException();
             return n1.x * n2.x + n1.y * n2.y + n1.z * n2.z;
         }
 
 
-        public static float AbsDot(Normal n, Vector v)
+        public static double AbsDot(Normal n, Vector v)
         {
             if (n.HasNaNs() || v.HasNaNs()) throw new InvalidOperationException();
             return Math.Abs(n.x * v.x + n.y * v.y + n.z * v.z);
         }
 
-        public static float AbsDot(Vector v, Normal n)
+        public static double AbsDot(Vector v, Normal n)
         {
             if (v.HasNaNs() || n.HasNaNs()) throw new InvalidOperationException();
             return Math.Abs(v.x * n.x + v.y * n.y + v.z * n.z);
         }
-        public static float AbsDot(Normal n1, Normal n2)
+        public static double AbsDot(Normal n1, Normal n2)
         {
             if (n1.HasNaNs() || n2.HasNaNs()) throw new InvalidOperationException();
             return Math.Abs(n1.x * n2.x + n1.y * n2.y + n1.z * n2.z);
         }
         public static Normal Faceforward(Normal n, Vector v)
         {
-            return (Dot(n, v) < 0.0f) ? -n : n;
+            return (Dot(n, v) < 0.0d) ? -n : n;
         }
         public static Normal Faceforward(Normal n1, Normal n2)
         {
-            return (Dot(n1, n2) < 0.0f) ? -n1 : n1;
+            return (Dot(n1, n2) < 0.0d) ? -n1 : n1;
         }
         public static Vector Faceforward(Vector v1, Vector v2)
         {
-            return (Dot(v1, v2) < 0.0f) ? -v1 : v1;
+            return (Dot(v1, v2) < 0.0d) ? -v1 : v1;
         }
         public static Vector Faceforward(Vector v, Normal n)
         {
-            return (Dot(v, n) < 0.0f) ? -v : v;
+            return (Dot(v, n) < 0.0d) ? -v : v;
         }
         public static BBox Union(BBox b, Point p)
         {
@@ -127,34 +127,34 @@ namespace nPBRT.Core.Geometry
             return ret;
         }
 
-        public static Vector SphericalDirection(float sintheta,
-                                 float costheta, float phi)
+        public static Vector SphericalDirection(double sintheta,
+                                 double costheta, double phi)
         {
-            return new Vector(sintheta * (float)Math.Cos(phi),
-                          sintheta * (float)Math.Sin(phi),
+            return new Vector(sintheta * Math.Cos(phi),
+                          sintheta * Math.Sin(phi),
                           costheta);
         }
 
 
-        public static Vector SphericalDirection(float sintheta, float costheta,
-                                         float phi, Vector x,
+        public static Vector SphericalDirection(double sintheta, double costheta,
+                                         double phi, Vector x,
                                            Vector y, Vector z)
         {
-            return sintheta * (float)Math.Cos(phi) * x +
-                   sintheta * (float)Math.Sin(phi) * y + costheta * z;
+            return sintheta * Math.Cos(phi) * x +
+                   sintheta * Math.Sin(phi) * y + costheta * z;
         }
 
 
-        public static float SphericalTheta(Vector v)
+        public static double SphericalTheta(Vector v)
         {
-            return (float)Math.Acos(Utility.Clamp(v.z, -1.0f, 1.0f));
+            return Math.Acos(Utility.Clamp(v.z, -1.0d, 1.0d));
         }
 
 
-        public static float SphericalPhi(Vector v)
+        public static double SphericalPhi(Vector v)
         {
-            float p = (float)Math.Atan2(v.y, v.x);
-            return (float)((p < 0.0f) ? p + 2.0f * Math.PI : p);
+            double p = Math.Atan2(v.y, v.x);
+            return ((p < 0.0d) ? p + 2.0d * Math.PI : p);
         }
 
     }
