+++ nPBRT/nPBRT/Core/Transform.cs
@@ -15,7 +15,7 @@ namespace nPBRT.Core
         {
 
         }
-        public Transform(float[,] mat)
+        public Transform(double[,] mat)
         {
             this.m = new Matrix4x4(mat[0, 0], mat[0, 1], mat[0, 2], mat[0, 3],
                           mat[1, 0], mat[1, 1], mat[1, 2], mat[1, 3],
@@ -43,48 +43,48 @@ namespace nPBRT.Core
         }
         public Point Apply(Point pt)
         {
-            float x = pt.x, y = pt.y, z = pt.z;
-            float xp = m.m[0, 0] * x + m.m[0, 1] * y + m.m[0, 2] * z + m.m[0, 3];
-            float yp = m.m[1, 0] * x + m.m[1, 1] * y + m.m[1, 2] * z + m.m[1, 3];
-            float zp = m.m[2, 0] * x + m.m[2, 1] * y + m.m[2, 2] * z + m.m[2, 3];
-            float wp = m.m[3, 0] * x + m.m[3, 1] * y + m.m[3, 2] * z + m.m[3, 3];
-            if (wp == 0.0f) throw new InvalidOperationException();
+            double x = pt.x, y = pt.y, z = pt.z;
+            double xp = m.m[0, 0] * x + m.m[0, 1] * y + m.m[0, 2] * z + m.m[0, 3];
+            double yp = m.m[1, 0] * x + m.m[1, 1] * y + m.m[1, 2] * z + m.m[1, 3];
+            double zp = m.m[2, 0] * x + m.m[2, 1] * y + m.m[2, 2] * z + m.m[2, 3];
+            double wp = m.m[3, 0] * x + m.m[3, 1] * y + m.m[3, 2] * z + m.m[3, 3];
+            if (wp == 0.0d) throw new InvalidOperationException();
             if (wp == 1.0) return new Point(xp, yp, zp);
             else return new Point(xp, yp, zp) / wp;
         }
         public void Apply(Point pt, ref Point ptrans)
         {
-            float x = pt.x, y = pt.y, z = pt.z;
+            double x = pt.x, y = pt.y, z = pt.z;
             ptrans.x = m.m[0, 0] * x + m.m[0, 1] * y + m.m[0, 2] * z + m.m[0, 3];
             ptrans.y = m.m[1, 0] * x + m.m[1, 1] * y + m.m[1, 2] * z + m.m[1, 3];
             ptrans.z = m.m[2, 0] * x + m.m[2, 1] * y + m.m[2, 2] * z + m.m[2, 3];
-            float w = m.m[3, 0] * x + m.m[3, 1] * y + m.m[3, 2] * z + m.m[3, 3];
-            if (w != 1.0f) ptrans /= w;
+            double w = m.m[3, 0] * x + m.m[3, 1] * y + m.m[3, 2] * z + m.m[3, 3];
+            if (w != 1.0d) ptrans /= w;
         }
         public Vector Apply(Vector v)
         {
-            float x = v.x, y = v.y, z = v.z;
+            double x = v.x, y = v.y, z = v.z;
             return new Vector(m.m[0, 0] * x + m.m[0, 1] * y + m.m[0, 2] * z,
                           m.m[1, 0] * x + m.m[1, 1] * y + m.m[1, 2] * z,
                           m.m[2, 0] * x + m.m[2, 1] * y + m.m[2, 2] * z);
         }
         public void Apply(Vector v, ref Vector vt)
         {
-            float x = v.x, y = v.y, z = v.z;
+            double x = v.x, y = v.y, z = v.z;
             vt.x = m.m[0, 0] * x + m.m[0, 1] * y + m.m[0, 2] * z;
             vt.y = m.m[1, 0] * x + m.m[1, 1] * y + m.m[1, 2] * z;
             vt.z = m.m[2, 0] * x + m.m[2, 1] * y + m.m[2, 2] * z;
         }
         public Normal Apply(Normal n)
         {
-            float x = n.x, y = n.y, z = n.z;
+            double x = n.x, y = n.y, z = n.z;
             return new Normal(mInv.m[0, 0] * x + mInv.m[1, 0] * y + mInv.m[2, 0] * z,
                           mInv.m[0, 1] * x + mInv.m[1, 1] * y + mInv.m[2, 1] * z,
                           mInv.m[0, 2] * x + mInv.m[1, 2] * y + mInv.m[2, 2] * z);
         }
         public void Apply(Normal n, ref Normal nt)
         {
-            float x = n.x, y = n.y, z = n.z;
+            double x = n.x, y = n.y, z = n.z;
             nt.x = mInv.m[0, 0] * x + mInv.m[1, 0] * y + mInv.m[2, 0] * z;
             nt.y = mInv.m[0, 1] * x + mInv.m[1, 1] * y + mInv.m[2, 1] * z;
             nt.z = mInv.m[0, 2] * x + mInv.m[1, 2] * y + mInv.m[2, 2] * z;
@@ -117,7 +117,7 @@ namespace nPBRT.Core
         }
         public bool SwapsHandedness()
         {
-            float det = ((m.m[0, 0] *
+            double det = ((m.m[0, 0] *
                           (m.m[1, 1] * m.m[2, 2] -
                            m.m[1, 2] * m.m[2, 1])) -
                           (m.m[0, 1] *
@@ -126,7 +126,7 @@ namespace nPBRT.Core
                           (m.m[0, 2] *
                           (m.m[1, 0] * m.m[2, 1] -
                            m.m[1, 1] * m.m[2, 0])));
-            return det < 0.0f;
+            return det < 0.0d;
         }
 
 
@@ -144,13 +144,13 @@ namespace nPBRT.Core
             int[] indxc = new int[4];
             int[] indxr = new int[4];
             int[] ipiv = { 0, 0, 0, 0 };
-            float[,] minv = new float[4, 4];
+            double[,] minv = new double[4, 4];
             Array.Copy(m.m, minv, 4 * 4);
 
             for (int i = 0; i < 4; i++)
             {
                 int irow = -1, icol = -1;
-                float big = 0.0f;
+                double big = 0.0d;
                 // Choose pivot
                 for (int j = 0; j < 4; j++)
                 {
@@ -162,7 +162,7 @@ namespace nPBRT.Core
                             {
                                 if (Math.Abs(minv[j, k]) >= big)
                                 {
-                                    big = (float)(Math.Abs(minv[j, k]));
+                                    big = (Math.Abs(minv[j, k]));
                                     irow = j;
                                     icol = k;
                                 }
@@ -177,16 +177,16 @@ namespace nPBRT.Core
                 if (irow != icol)
                 {
                     for (int k = 0; k < 4; ++k)
-                        Utility.Swap<float>(ref minv[irow, k], ref  minv[icol, k]);
+                        Utility.Swap<double>(ref minv[irow, k], ref  minv[icol, k]);
                 }
                 indxr[i] = irow;
                 indxc[i] = icol;
-                if (minv[icol, icol] == 0.0f)
+                if (minv[icol, icol] == 0.0d)
                     throw new Exception("Singular matrix in MatrixInvert");
 
                 // Set $m[icol][icol]$ to one by scaling row _icol_ appropriately
-                float pivinv = 1.0f / minv[icol, icol];
-                minv[icol, icol] = 1.0f;
+                double pivinv = 1.0d / minv[icol, icol];
+                minv[icol, icol] = 1.0d;
                 for (int j = 0; j < 4; j++)
                     minv[icol, j] *= pivinv;
 
@@ -195,7 +195,7 @@ namespace nPBRT.Core
                 {
                     if (j != icol)
                     {
-                        float save = minv[j, icol];
+                        double save = minv[j, icol];
                         minv[j, icol] = 0;
                         for (int k = 0; k < 4; k++)
                             minv[j, k] -= minv[icol, k] * save;
@@ -208,7 +208,7 @@ namespace nPBRT.Core
                 if (indxr[j] != indxc[j])
                 {
                     for (int k = 0; k < 4; k++)
-                        Utility.Swap<float>(ref minv[k, indxr[j]], ref minv[k, indxc[j]]);
+                        Utility.Swap<double>(ref minv[k, indxr[j]], ref minv[k, indxc[j]]);
                 }
             }
             return new Matrix4x4(minv);
@@ -228,24 +228,24 @@ namespace nPBRT.Core
         }
 
 
-        public static Transform Scale(float x, float y, float z)
+        public static Transform Scale(double x, double y, double z)
         {
             Matrix4x4 m = new Matrix4x4(x, 0, 0, 0,
                         0, y, 0, 0,
                         0, 0, z, 0,
                         0, 0, 0, 1);
-            Matrix4x4 minv = new Matrix4x4(1.0f / x, 0, 0, 0,
-                           0, 1.0f / y, 0, 0,
-                           0, 0, 1.0f / z, 0,
+            Matrix4x4 minv = new Matrix4x4(1.0d / x, 0, 0, 0,
+                           0, 1.0d / y, 0, 0,
+                           0, 0, 1.0d / z, 0,
                            0, 0, 0, 1);
             return new Transform(m, minv);
         }
 
 
-        public static Transform RotateX(float angle)
+        public static Transform RotateX(double angle)
         {
-            float sin_t = (float)Math.Sin(Utility.Radians(angle));
-            float cos_t = (float)Math.Cos(Utility.Radians(angle));
+            double sin_t = Math.Sin(Utility.Radians(angle));
+            double cos_t = Math.Cos(Utility.Radians(angle));
             Matrix4x4 m = new Matrix4x4(1, 0, 0, 0,
                         0, cos_t, -sin_t, 0,
                         0, sin_t, cos_t, 0,
@@ -254,10 +254,10 @@ namespace nPBRT.Core
         }
 
 
-        public static Transform RotateY(float angle)
+        public static Transform RotateY(double angle)
         {
-            float sin_t = (float)Math.Sin(Utility.Radians(angle));
-            float cos_t = (float)Math.Cos(Utility.Radians(angle));
+            double sin_t = Math.Sin(Utility.Radians(angle));
+            double cos_t = Math.Cos(Utility.Radians(angle));
             Matrix4x4 m = new Matrix4x4(cos_t, 0, sin_t, 0,
                          0, 1, 0, 0,
                         -sin_t, 0, cos_t, 0,
@@ -267,11 +267,11 @@ namespace nPBRT.Core
 
 
 
-        public static Transform RotateZ(float angle)
+        public static Transform RotateZ(double angle)
         {
 
-            float sin_t = (float)Math.Sin(Utility.Radians(angle));
-            float cos_t = (float)Math.Cos(Utility.Radians(angle));
+            double sin_t = Math.Sin(Utility.Radians(angle));
+            double cos_t = Math.Cos(Utility.Radians(angle));
             Matrix4x4 m = new Matrix4x4(cos_t, -sin_t, 0, 0,
                         sin_t, cos_t, 0, 0,
                         0, 0, 1, 0,
@@ -280,26 +280,26 @@ namespace nPBRT.Core
         }
 
 
-        public static Transform Rotate(float angle, Vector axis)
+        public static Transform Rotate(double angle, Vector axis)
         {
             Vector a = Geometry.Geometry.Normalize(axis);
-            float s = (float)Math.Sin(Utility.Radians(angle));
-            float c = (float)Math.Cos(Utility.Radians(angle));
-            float[,] m = new float[4, 4];
+            double s = Math.Sin(Utility.Radians(angle));
+            double c = Math.Cos(Utility.Radians(angle));
+            double[,] m = new double[4, 4];
 
-            m[0, 0] = a.x * a.x + (1.0f - a.x * a.x) * c;
-            m[0, 1] = a.x * a.y * (1.0f - c) - a.z * s;
-            m[0, 2] = a.x * a.z * (1.0f - c) + a.y * s;
+            m[0, 0] = a.x * a.x + (1.0d - a.x * a.x) * c;
+            m[0, 1] = a.x * a.y * (1.0d - c) - a.z * s;
+            m[0, 2] = a.x * a.z * (1.0d - c) + a.y * s;
             m[0, 3] = 0;
 
-            m[1, 0] = a.x * a.y * (1.0f - c) + a.z * s;
-            m[1, 1] = a.y * a.y + (1.0f - a.y * a.y) * c;
-            m[1, 2] = a.y * a.z * (1.0f - c) - a.x * s;
+            m[1, 0] = a.x * a.y * (1.0d - c) + a.z * s;
+            m[1, 1] = a.y * a.y + (1.0d - a.y * a.y) * c;
+            m[1, 2] = a.y * a.z * (1.0d - c) - a.x * s;
             m[1, 3] = 0;
 
-            m[2, 0] = a.x * a.z * (1.0f - c) - a.y * s;
-            m[2, 1] = a.y * a.z * (1.0f - c) + a.x * s;
-            m[2, 2] = a.z * a.z + (1.0f - a.z * a.z) * c;
+            m[2, 0] = a.x * a.z * (1.0d - c) - a.y * s;
+            m[2, 1] = a.y * a.z * (1.0d - c) + a.x * s;
+            m[2, 2] = a.z * a.z + (1.0d - a.z * a.z) * c;
             m[2, 3] = 0;
 
             m[3, 0] = 0;
@@ -314,7 +314,7 @@ namespace nPBRT.Core
 
         public static Transform LookAt(Point pos, Point look, Vector up)
         {
-            float[,] m = new float[4, 4];
+            double[,] m = new double[4, 4];
             // Initialize fourth column of viewing matrix
             m[0, 3] = pos.x;
             m[1, 3] = pos.y;
@@ -328,15 +328,15 @@ namespace nPBRT.Core
             m[0, 0] = left.x;
             m[1, 0] = left.y;
             m[2, 0] = left.z;
-            m[3, 0] = 0.0f;
+            m[3, 0] = 0.0d;
             m[0, 1] = newUp.x;
             m[1, 1] = newUp.y;
             m[2, 1] = newUp.z;
-            m[3, 1] = 0.0f;
+            m[3, 1] = 0.0d;
             m[0, 2] = dir.x;
             m[1, 2] = dir.y;
             m[2, 2] = dir.z;
-            m[3, 2] = 0.0f;
+            m[3, 2] = 0.0d;
             Matrix4x4 camToWorld = new Matrix4x4(m);
             return new Transform(Inverse(camToWorld), camToWorld);
         }
